requirement.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
bs4
requests
validators
youtube-transcript-api
pytube
langchain-text-splitters